revised the implementation of indexing trees and weak reference tables
This change fixes major problems of the previous implementation in
(java.rt.map.*):
 * excessively redundant code
 * a lot of casting expressions, which could be avoided
 * exposing unsupported interfaces
 * confusing names
 * inconsistent operations, which look very suspicious

The new implementation eliminates these problems; especially, it relies
on type system, so that the compiler can detect prohibited operations.
This also completely eliminates casting.

This change merges a lot of scattered code that was slightly
inconsistent; I picked the most reasonable one. Hopefully, this fixes
the problem that might be caused by inconsistency.