added no weak-reference interning option
Since weak-reference interning, introduced in JavaMOP 3.0, can degrade performance, it was necessary to allow the user to turn off this.

When interning is disabled, the GWRT is not used. As a result, multiple weak references for the same object can be created.

This change does not affect the runtime library.