implemented new way to generate code while fixing the bug that wrongly skips events
Now, the generated code is type-safe. It does not contain type-casting, unless it is definitely needed. One such example is when an entry can contain either a fully-fledged monitor or a 'disable' holder. Other than that, there should be no type-casting, which enables the compiler to type-check rigorously.

To do generate type-safe code and fix the bug, the way to generate code has been changed a lot. Previously, generating code was done by string concatenation, which is hard to maintain, in my opinion, and hard to analyze. This changelist generates code by creating CodeDOM. Each expression and statement is represented as a CodeDOM object, and code is generated by iterating over such objects. Most of all, this enables code analysis, including dead-code elimination, which can possibly improves performance.

Additionally, the '-internalbehavior' option has been added. If this is enabled, the generated class implements the IObservable interface, which allows any observer to listen to internal events, such as "what indexing trees are looked up",  "what monitors are created", "which events are dispatched to what monitors" and so forth. This is used to test the behavior thoroughly.

This changelist affects mostly 'event method'-generating code and indexing trees. (By 'event method', I meant the static method exposed to clients.)

Here's the bug description:
Since JavaMOP 3.0, 'disable' and 't' have been stored at the weak reference. Due to weak-refernce interning, a single weak reference is created for the same object and, consequently, monitors with different parameter bindings share the same 'disable' and 't' values, which is incorrect.