implemented the code simplifier
The implemented code simplifier eliminates a variable that is created (and optionally assigned) but never used. As a result, the generated code has no warnings caused by "unused variables" and may help the Java and JIT compilers to generate better code.

This changelist also fixes the null pointer exception raised when an event with no parameters occurs.